Fix new warning when compiling with c++98

+++ src/entropy/BinaryEntropyEncoder.cpp
@@ -27,7 +27,7 @@ const uint64 BinaryEntropyEncoder::MASK_0_24 = 0x0000000000FFFFFF;
 const uint64 BinaryEntropyEncoder::MASK_0_32 = 0x00000000FFFFFFFF;
 const int BinaryEntropyEncoder::MAX_BLOCK_SIZE = 1 << 30;
 const int BinaryEntropyEncoder::MAX_CHUNK_SIZE = 1 << 26;
-static const uint64 BINARY_ENTROPY_BUFFER_FLOOR = 8ULL << 20;
+static const uint64 BINARY_ENTROPY_BUFFER_FLOOR = uint64(8) << 20;
 
 
 BinaryEntropyEncoder::BinaryEntropyEncoder(OutputBitStream& bitstream, Predictor* predictor, bool deallocate)
